Simplify component structure, by using layouts
